+++ karma.conf.js
@@ -70,7 +70,7 @@ module.exports = function (config) {
             {pattern: "node_modules/angular2-text-mask/**/*.js", included: false, watched: false},
             {pattern: "node_modules/text-mask-core/**/*.js", included: false, watched: false},
 
-            {pattern: "node_modules/ionic-angular/**/*.js", included: false, watched: false},
+            {pattern: "node_modules/@ionic/angular/**/*.js", included: false, watched: false},
 
             {pattern: "node_modules/ngx-bootstrap/**/*.js", included: false, watched: false},
 
@@ -147,6 +147,12 @@ module.exports = function (config) {
 
         // Concurrency level
         // how many browser should be started simultaneous
-        concurrency: Infinity
+        concurrency: Infinity,
+
+        client: {
+            jasmine: {
+                random: false
+            }
+        }
     })
 };