+++ packages/ui-ionic/src/datetime/dynamic-ionic-datetime.component.spec.ts
@@ -4,7 +4,7 @@ import { ReactiveFormsModule, FormGroup, FormControl } from "@angular/forms";
 import { NoopAnimationsModule } from "@angular/platform-browser/animations";
 import { By } from "@angular/platform-browser";
 import { TextMaskModule } from "angular2-text-mask";
-import { DateTime, IonicModule } from "ionic-angular";
+import { Datetime, IonicModule } from "@ionic/angular";
 import { DynamicDatePickerModel, DynamicFormsCoreModule, DynamicFormService } from "@ng-dynamic-forms/core";
 import { DynamicIonicDateTimeComponent } from "./dynamic-ionic-datetime.component";
 
@@ -58,7 +58,7 @@ xdescribe("DynamicIonicDateTimeComponent test suite", () => {
         expect(component.control instanceof FormControl).toBe(true);
         expect(component.group instanceof FormGroup).toBe(true);
         expect(component.model instanceof DynamicDatePickerModel).toBe(true);
-        expect(component.ionDateTime instanceof DateTime).toBe(true);
+        expect(component.ionDatetime instanceof Datetime).toBe(true);
 
         expect(component.blur).toBeDefined();
         expect(component.change).toBeDefined();