+++ packages/ui-material/src/chips/dynamic-material-chips.component.ts
@@ -1,13 +1,4 @@
-import {
-    ChangeDetectionStrategy,
-    Component,
-    EventEmitter,
-    Inject,
-    Input,
-    Optional,
-    Output,
-    ViewChild
-} from "@angular/core";
+import { Component, EventEmitter, Inject, Input, OnDestroy, Optional, Output, ViewChild } from "@angular/core";
 import { FormGroup } from "@angular/forms";
 import {
     LabelOptions,
@@ -18,24 +9,25 @@ import {
     MatChipsDefaultOptions,
     MatInput
 } from "@angular/material";
+import { Subscription } from "rxjs";
 import {
+    DynamicFormControlComponent,
     DynamicFormControlCustomEvent,
     DynamicFormLayout,
     DynamicFormLayoutService,
     DynamicFormValidationService,
-    DynamicFormControlComponent,
     DynamicInputModel
 } from "@ng-dynamic-forms/core";
 
 @Component({
     selector: "dynamic-material-chips",
-    templateUrl: "./dynamic-material-chips.component.html",
-    changeDetection: ChangeDetectionStrategy.OnPush
+    templateUrl: "./dynamic-material-chips.component.html"
 })
-export class DynamicMaterialChipsComponent extends DynamicFormControlComponent {
+export class DynamicMaterialChipsComponent extends DynamicFormControlComponent implements OnDestroy {
 
     private _chipList: string[];
     private _model: DynamicInputModel;
+    private _valueSubscription: Subscription;
 
     @Input() bindId: boolean = true;
     @Input() group: FormGroup;
@@ -47,8 +39,13 @@ export class DynamicMaterialChipsComponent extends DynamicFormControlComponent {
     }
 
     set model(model: DynamicInputModel) {
+
+        this.unsubscribe();
+
         this._model = model;
-        this._chipList = model.value as string[] || [];
+        this._model.valueUpdates.subscribe((value: string[]) => this.chipList = value);
+
+        this.chipList = Array.isArray(model.value) ? model.value as string[] : [];
     }
 
     @Output() blur: EventEmitter<any> = new EventEmitter();
@@ -67,8 +64,22 @@ export class DynamicMaterialChipsComponent extends DynamicFormControlComponent {
         super(layoutService, validationService);
     }
 
+    ngOnDestroy() {
+        this.unsubscribe();
+    }
+
+    unsubscribe(): void {
+        if (this._valueSubscription) {
+            this._valueSubscription.unsubscribe();
+        }
+    }
+
     get chipList(): string[] {
         return this._chipList;
+    }
+
+    set chipList(value: string[]) {
+        this._chipList = value;
     }
 
     onChipInputTokenEnd($event: MatChipInputEvent): void {